:recycle: Refactor: ref_link 예외 및 최대 인원 관련 예외 처리 추가
1. 추가한 기능
- ref_link 예외 처리
    - link_type만 선택되고, url이 작성되지 않았을 경우
    - link_type이 선택되지 않고, url이 작성되었을 경우
    - ref_link의 모든 요소가 작성되지 않았을 경우
- max_members 관련 처리
    - template 내에서 min과 max 값을 지정하는 것으로 4명 이상 20명 이하
      처리
    - 추가적인 처리가 필요하다면 서버 쪽에서 처리하겠으나, 현재는
      클라이언트에서만 처리해도 문제없을 것으로 판단됨.
2. 비고
- thumbnail 이미지 업로드 시 default 이미지가 저장되지 않는 상태 발생

+++ studies/views.py
@@ -159,14 +159,15 @@ class StudyCreate(LoginRequiredMixin, CreateView):
             )
 
         ref_links_input = form.cleaned_data["ref_links"]
-        ref_links_list = ref_links_input.strip(",").split(",")
+        if ref_links_input != "":
+            ref_links_list = ref_links_input.strip(",").split(",")
 
-        for ref_link in ref_links_list:
-            RefLink.objects.create(
-                link_type=ref_link.split(";")[0].strip(),
-                url=ref_link.split(";")[1].strip(),
-                study=study,
-            )
+            for ref_link in ref_links_list:
+                RefLink.objects.create(
+                    link_type=ref_link.split(";")[0].strip(),
+                    url=ref_link.split(";")[1].strip(),
+                    study=study,
+                )
 
         return super().form_valid(form)
 
@@ -238,16 +239,16 @@ class StudyUpdate(UserPassesTestMixin, UpdateView):
             )
 
         ref_links_input = form.cleaned_data["ref_links"]
-        ref_links_list = ref_links_input.strip(",").split(",")
+        if ref_links_input != "":
+            RefLink.objects.filter(study=study).delete()
+            ref_links_list = ref_links_input.strip(",").split(",")
 
-        RefLink.objects.filter(study=study).delete()
-
-        for ref_link in ref_links_list:
-            RefLink.objects.create(
-                link_type=ref_link.split(";")[0].strip(),
-                url=ref_link.split(";")[1].strip(),
-                study=study,
-            )
+            for ref_link in ref_links_list:
+                RefLink.objects.create(
+                    link_type=ref_link.split(";")[0].strip(),
+                    url=ref_link.split(";")[1].strip(),
+                    study=study,
+                )
 
         return super().form_valid(form)
